backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
databases==0.8.0
psycopg2-binary==2.9.9
openai==1.3.7
pydantic==2.5.0
python-multipart==0.0.6